famfs: Add a shadowfs option to logplay command
Create a shadow filesystem of Famfs filesystem.
This feature adds a -S option which takes 2 arguments
1. Path to an existing directory in the filesystem
2. Path to a mounted famfs filessytem.
This option replays the log and recreates the famfs
files and directory structure in the path pointed by arg 1.
For each file in the famfs filesystem a corresponding file
entry is created in the shadow filesystem and these files
contain the metadata of the original files.

There is also an added function to get the metadata from the
files in the shadow filesystem. The function famfs_get_shadow_file_data()
reads the metadata from the file in shadowfs and returns it in the form of
struct famfs_file_creation*. This can be modified in future.

Usage:
./famfs logplay -vv -S /tmp/famfsshadow/ /mnt/famfs/

Signed-off-by: Aravind Ramesh <arramesh@micron.com>