famfs_fused: file descriptor conservation, and new flock implementation
Previously all famfs file opens corresponded to an open on the
corresponding shadow yaml file. This was kinda dumb, and wasted a lot of
file descriptors with some workloads. Instead, now each open gets a ref
on the famfs_inode, and release(close) puts the ref.

Not keeping shadow yaml files open also broke flock(), because we were
just passing flock through to the backing shadow yaml file. I've done a
new implementation of flock, as follows:

* flock takes uses icache->flock_mutex as follows:
  * flock(LOCK_EX) locks the mutex
  * flock(LOCK_UN) unlocks the mutex
  * inode->flock_held remembers whether the file is flocked
  * famfs_release unlocks the mutex if (inode->flock_held)

Note that there is only one icache->flock_mutex, so a flock on any file
will make all files look locked. I'm ok with this because I'm only
enabling flock so that the famfs user space can flock(.meta/.log); this
is how we serialize file allocation and log append.

So our policy is flock is only supported for the famfs user space code,
and not for user applications.

If we wanted a ful general-purpose flock implementation, we could
replace icache->flock_mutex with an inode->flock_mutex - in every inode.
Or maybe a hashed lock pool. But that doesn't seem worth it now since
there is only one known flock user, and it's always the same file
(.meta/.log).